requirements.txt
pip
setuptools
numpy
discord.py[voice]
google-api-python-client
google-cloud-storage
sentry-sdk
pydantic